requirements.txt
crewai
crewai_tools
python-dotenv
langchain-groq
browserbase